f/postAlgorithm
