+++ client/components/home-page/index.js
@@ -15,7 +15,6 @@ class HomePage extends Component {
 
     showQuestions(event) {
         event.preventDefault()
-        console.log(event.target.getAttribute("value"))
         this.setState({showQuestions: event.target.getAttribute("value")})
     }
