+++ client/store/index.js
@@ -3,12 +3,12 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import user from './user'
-import userInput from './user-input'
+import algorithmInput from './algorithm-input'
 import questions from './questions'
 import categories from './categories'
 
 
-const reducer = combineReducers({ user, userInput, questions, categories})
+const reducer = combineReducers({ user, algorithmInput, questions, categories })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
 )
@@ -16,6 +16,6 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
-export * from './user-input'
+export * from './algorithm-input'
 export * from './questions'
 export * from './categories'
